improving code and file structure
